requirements.txt
langchain<0.2.0
langchain_huggingface
pymupdf
unstructured
python-docx
flask
pypdf
python-dotenv
langchain_community
faiss-cpu
sentence-transformers==2.7.0
huggingface_hub==0.23.0
transformers==4.41.2
accelerate==0.30.1
bitsandbytes
peft
trl
PyYAML
fastapi==0.115.12
uvicorn==0.34.1
ipykernel